% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{sf_linestring}
\alias{sf_linestring}
\title{sf LINESTRING}
\usage{
sf_linestring(
  obj = NULL,
  x = NULL,
  y = NULL,
  z = NULL,
  m = NULL,
  linestring_id = NULL,
  keep = FALSE,
  list_columns = NULL
)
}
\arguments{
\item{obj}{sorted matrix or data.frame}

\item{x}{x geometry column}

\item{y}{y geometry column}

\item{z}{z geometry column}

\item{m}{m geometry column}

\item{linestring_id}{column of ids for linestrings}

\item{keep}{logical indicating if the non-geometry and non-id columns should be kept.
if TRUE you must supply the geometry and id columns, and only the first row of
each geometry is kept. See Keeping Properties.}

\item{list_columns}{vector of column names to turn into a list.}
}
\value{
\code{sf} object of LINESTRING geometries
}
\description{
constructs sf of LINESTRING objects
}
\examples{

x <- matrix( c(1:4), ncol = 2 )
sf_linestring( x )

x <- data.frame( id = 1:2, x = 1:2, y = 2:1 )
sf_linestring( x )
sf_linestring( x, x = "x", y = "y" )
sf_linestring( x, x = "y", y = "x" )
sf_linestring( x, linestring_id = "id", x = "x", y = "y")

## keeping properties
x <- data.frame( id = c(1,1,2,2), x = 1:4, y = 4:1, val = letters[1:4], stringsAsFactors = FALSE)

## first-row of 'val' is kept
sf_linestring( x, linestring_id = "id", x = "x", y = "y", keep = TRUE )

## 'val' column converted to a list
sf_linestring( x, linestring_id = "id", x = "x", y = "y", keep = TRUE, list_columns = "val" )


}
