% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{sf_multipoint}
\alias{sf_multipoint}
\title{sf MULTIPOINT}
\usage{
sf_multipoint(
  obj,
  x = NULL,
  y = NULL,
  z = NULL,
  m = NULL,
  multipoint_id = NULL,
  keep = FALSE,
  list_columns = NULL
)
}
\arguments{
\item{obj}{sorted matrix or data.frame}

\item{x}{x geometry column}

\item{y}{y geometry column}

\item{z}{z geometry column}

\item{m}{m geometry column}

\item{multipoint_id}{column of ids for multipoints}

\item{keep}{logical indicating if the non-geometry and non-id columns should be kept.
if TRUE you must supply the geometry and id columns, and only the first row of
each geometry is kept. See Keeping Properties.}

\item{list_columns}{vector of column names to turn into a list.}
}
\value{
\code{sf} object of MULTIPOINT geometries
}
\description{
constructs sf of MULTIPOINT objects
}
\examples{

x <- matrix( c(1:4), ncol = 2 )
sf_multipoint( x )

x <- data.frame( id = 1:2, x = 1:2, y = 2:1 )
sf_multipoint( x )
sf_multipoint( x, x = "x", y = "y" )
sf_multipoint( x, x = "y", y = "x" )
sf_multipoint( x, multipoint_id = "id", x = "x", y = "y")

}
