% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc.R
\name{sfc_multilinestring}
\alias{sfc_multilinestring}
\title{sfc MULTILINESTRING}
\usage{
sfc_multilinestring(obj, x = NULL, y = NULL, z = NULL, m = NULL,
  multilinestring_id = NULL, linestring_id = NULL)
}
\description{
sfc MULTILINESTRING
}
\examples{

m <- matrix(c(0,0,0,0,1,1), ncol = 3 )
sfc_multilinestring( m )

m <- matrix(c(0,0,0,0,0,1,0,1,1,1,2,2,1,2,3), ncol = 3, byrow = T)
sfc_multilinestring( obj = m,  multilinestring_id = 1 )



}
