% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfc.R
\name{sfc_point}
\alias{sfc_point}
\title{sfc POINTS}
\usage{
sfc_point(obj, x = NULL, y = NULL, z = NULL, m = NULL)
}
\arguments{
\item{obj}{matrix or data.frame}

\item{x}{x geometry column}

\item{y}{y geometry column}

\item{z}{z geometry column}

\item{m}{m geometry column}
}
\description{
constructs sfc of POINT objects
}
\examples{

x <- c(1:3)
sfc_point( x )

x <- matrix( c(1:10) , ncol = 2 )
sfc_point( x )

x <- setNames( as.data.frame( x ), c("x","y") )
sfc_point( x )
sfc_point( obj = x, x = "x", y = "y" )
sfc_point( obj = x, x = "y", y = "x" )

}
