% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\name{sfc_to_df}
\alias{sfc_to_df}
\title{sfc to df}
\usage{
sfc_to_df(sfc)
}
\arguments{
\item{sfc}{sfc object}
}
\description{
Converts an sfc object to a data.frame
}
\examples{

x <- matrix( c(1:16), ncol = 2 )
sfc <- sfc_linestring( x )
df <- sfc_to_df( sfc )

df <- data.frame(
ml_id = c(1,1,1,1,1,1,1,1,2,2,2,2,2)
, l_id = c(1,1,1,2,2,3,3,3,1,1,1,2,2)
, x = rnorm(13)
, y = rnorm(13)
, z = rnorm(13)
, m = rnorm(13)
)
sfc <- sfc_multilinestring( obj = df, multilinestring_id = "ml_id", linestring_id = "l_id" )

df <- sfc_to_df( sfc )

}
