% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfg.R
\name{sfg_polygon}
\alias{sfg_polygon}
\title{to MULTILINESTRING}
\usage{
sfg_polygon(obj, x = NULL, y = NULL, z = NULL, m = NULL,
  line_id = NULL)
}
\description{
to MULTILINESTRING
}
\examples{

sfg_polygon( matrix( 1:24, ncol = 2 ) )
sfg_polygon( matrix( 1:24, ncol = 3 ) )
sfg_polygon( matrix( 1:24, ncol = 4 ) )

## different lines
m <- cbind( matrix( 1:24, ncol = 2 ), c(rep(1, 6), rep(2, 6) ) )
sfg_polygon( obj = m, x = 1, y = 2, line_id = 3 )

## just specifying line_id will use all others as the geometries
sfg_polygon( obj = m, line_id = 3 )

df <- data.frame( x = 1:12, y = 1:12, z = 13:24, id = c(rep(1,6), rep(2,6)))
sfg_polygon( df, x = "x", y = "y" )
sfg_polygon( df, x = "x", y = "y", line_id = "id" )

sfg_polygon( df, line_id = "id" )


}
