% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\name{sfg_to_df}
\alias{sfg_to_df}
\title{sfg to df}
\usage{
sfg_to_df(sfg)
}
\arguments{
\item{sfg}{sfg object}
}
\description{
Converts an sfg object to a data.frame
}
\examples{
sfg <- sfg_point( obj = c(1,2) )
df <- sfg_to_df( sfg )

m <- cbind( matrix( 1:24, ncol = 2 ), c(rep(1, 6), rep(2, 6) ) )
sfg <- sfg_polygon( obj = m, x = 1, y = 2, linestring_id = 3 )
df <- sfg_to_df( sfg )


}
