% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_sfg.R
\name{to_multipoint}
\alias{to_multipoint}
\title{to MULTIPOINT}
\usage{
to_multipoint(obj, x = NULL, y = NULL, z = NULL, m = NULL)
}
\description{
to MULTIPOINT
}
\examples{

to_multipoint( 1:2 )
to_multipoint( 1:3 )
to_multipoint( 1:4 )

to_multipoint( matrix( 1:3, ncol = 3 ) )
to_multipoint( data.frame( x = 1, y = 2, z = 3 )  )

to_multipoint( matrix( 1:4, ncol = 2 ) )
to_multipoint( matrix( 1:24, ncol = 2, byrow = T ) )
to_multipoint( matrix( 1:24, ncol = 3, byrow = T ) )
to_multipoint( matrix( 1:24, ncol = 4, byrow = T ) )

to_multipoint( data.frame( x = 1:5, y = 1:5 ) )

## using columns

to_multipoint( matrix( 1:24, ncol = 4, byrow = T ), x = 1, y = 2 )
to_multipoint( matrix( 1:24, ncol = 4, byrow = T ), x = 1, y = 2, z = 3 )
to_multipoint( matrix( 1:24, ncol = 4, byrow = T ), x = 3, y = 4 )

to_multipoint( data.frame( x = 1:5, y = 1:5, z = 11:15, m = 11:15 ), x = "x", y = "y" )
to_multipoint( data.frame( x = 1:5, y = 1:5, z = 11:15, m = 11:15 ), x = "x", y = "y", z = "z" )
to_multipoint( data.frame( x = 1:5, y = 1:5, z = 11:15, m = 11:15 ), x = "x", y = "y", z = "z", m = "m" )

}
