% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_sfc.R
\name{to_multipoints}
\alias{to_multipoints}
\title{to MULTIPOINTS}
\usage{
to_multipoints(obj, x = NULL, y = NULL, z = NULL, m = NULL,
  multipoint_id = NULL)
}
\description{
constructs sfc of MULTIPOINT objects
}
\examples{

x <- matrix( c(1:4), ncol = 2 )
to_multipoints( x )

x <- data.frame( id = 1:2, x = 1:2, y = 2:1 )
to_multipoints( x )
to_multipoints( x, x = "x", y = "y" )
to_multipoints( x, x = "y", y = "x" )
to_multipoints( x, multipoint_id = "id", x = "x", y = "y")

}
