% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_sfg.R
\name{to_multipolygon}
\alias{to_multipolygon}
\title{to MULTIPOLYGON}
\usage{
to_multipolygon(obj, x = NULL, y = NULL, z = NULL, m = NULL,
  polygon_id = NULL, line_id = NULL)
}
\description{
to MULTIPOLYGON
}
\examples{

df <- data.frame(
  polygon_id = c(rep(1, 5), rep(2, 10))
  , line_id = c(rep(1, 10), rep(2, 5))
  , x = c(0,0,1,1,0,2,2,5,5,2,3,3,4,4,3)
  , y = c(0,1,1,0,0,2,5,5,2,2,3,4,4,3,3)
  , z = c(1)
  , m = c(1)
)

m <- as.matrix( df )

to_multipolygon( df[, c("x","y") ] )

to_multipolygon( df, x = "x", y = "y", polygon_id = "polygon_id", line_id = "line_id" )
to_multipolygon( df, x = "x", y = "y", z = "z", polygon_id = "polygon_id", line_id = "line_id" )
to_multipolygon( df, x = "x", y = "y", z = "z", m = "m", polygon_id = "polygon_id", line_id = "line_id" )


to_multipolygon( m[, c("x","y") ] )

to_multipolygon( m, x = "x", y = "y", polygon_id = "polygon_id", line_id = "line_id" )
to_multipolygon( m, x = "x", y = "y", z = "z", polygon_id = "polygon_id", line_id = "line_id" )
to_multipolygon( m, x = "x", y = "y", z = "z", m = "m", polygon_id = "polygon_id", line_id = "line_id" )

}
