% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_sfg.R
\name{to_point}
\alias{to_point}
\title{to POINT}
\usage{
to_point(obj, x = NULL, y = NULL, z = NULL, m = NULL)
}
\description{
to POINT
}
\examples{

to_point( 1:2 )
to_point( 1:3 )
to_point( 1:4 )

to_point( matrix( 1:3, ncol = 3 ) )
to_point( data.frame( x = 1, y = 2, z = 3 )  )

to_point( data.frame( x = 1, y = 2, z = 3 ), x = "x", y = "y" )
to_point( data.frame( x = 1, y = 2, z = 3 ), x = 1, y = 3  )

}
