% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_sfc.R
\name{to_points}
\alias{to_points}
\title{to POINTS}
\usage{
to_points(obj, x = NULL, y = NULL, z = NULL, m = NULL)
}
\description{
constructs sfc of POINT objects
}
\examples{

x <- c(1:3)
to_points( x )

x <- matrix( c(1:10) , ncol = 2 )
to_points( x )

x <- setNames( as.data.frame( x ), c("x","y") )
to_points( x )
to_points( obj = x, x = "x", y = "y" )
to_points( obj = x, x = "y", y = "x" )

}
