% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_sfc.R
\name{to_sfc}
\alias{to_sfc}
\title{to sfc}
\usage{
to_sfc(obj, x, y, z = NULL, m = NULL, point_id = NULL,
  multipoint_id = NULL, linestring_id = NULL,
  multilinestring_id = NULL, polygon_id = NULL,
  multipolygon_id = NULL)
}
\description{
converts a data.frame to an sfc object
}
