% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_sfg.R
\name{to_sfg}
\alias{to_sfg}
\title{to sfc}
\usage{
to_sfg(obj, x, y, z = NULL, m = NULL, point_id = NULL,
  multipoint_id = NULL, linestring_id = NULL,
  multilinestring_id = NULL, polygon_id = NULL,
  multipolygon_id = NULL)
}
\description{
converts a data.frame to an sfc object
}
\examples{

df <- data.frame(
  id = rep(1,5),
  x = 1:5,
  y = 6:10,
  z = 11:15
)

}
