% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfg.R
\name{sfg_linestring}
\alias{sfg_linestring}
\title{sfg linestring}
\usage{
sfg_linestring(obj, x = NULL, y = NULL, z = NULL, m = NULL)
}
\arguments{
\item{obj}{matrix or data.frame}

\item{x}{x geometry column}

\item{y}{y geometry column}

\item{z}{z geometry column}

\item{m}{m geometry column}
}
\value{
\code{sfg} object of LINESTRING geometry
}
\description{
constructs sfg LINESTRING object
}
\examples{

sfg_linestring( matrix( 1:24, ncol = 2 ) )
sfg_linestring( matrix( 1:24, ncol = 3 ) )
sfg_linestring( matrix( 1:24, ncol = 4 ) )

sfg_linestring( matrix( 1:24, ncol = 4 ), x = 3, y = 2, z = 3)

sfg_linestring( data.frame( x = 1:10, y = 11:20 ) )
sfg_linestring( data.frame( x = 1:10, y = 11:20, z = 21:30 ) )
sfg_linestring( data.frame( x = 1:10, y = 11:20, z = 21:30 ), x = "x", y = "z" )


}
