% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfg.R
\name{sfg_point}
\alias{sfg_point}
\title{sfg point}
\usage{
sfg_point(obj, x = NULL, y = NULL, z = NULL, m = NULL)
}
\arguments{
\item{obj}{matrix or data.frame}

\item{x}{x geometry column}

\item{y}{y geometry column}

\item{z}{z geometry column}

\item{m}{m geometry column}
}
\value{
\code{sfg} object of POINT geometry
}
\description{
constructs sfg POINT object
}
\examples{

sfg_point( 1:2 )
sfg_point( 1:3 )
sfg_point( 1:4 )

sfg_point( matrix( 1:3, ncol = 3 ) )
sfg_point( data.frame( x = 1, y = 2, z = 3 )  )

sfg_point( data.frame( x = 1, y = 2, z = 3 ), x = "x", y = "y" )
sfg_point( data.frame( x = 1, y = 2, z = 3 ), x = 1, y = 3  )

}
